office_app/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
bcrypt==4.1.2
python-dotenv==1.0.0

